+++ src/cargo.ts
@@ -31,9 +31,12 @@ export type ContractNames =
   | 'cargoSell'
   | 'cargoData'
   | 'cargoFunds'
+  | 'cargoBatchMint'
+  | 'cargoTokenV1Creator'
+  | 'cargoTokenV2Creator'
+  | 'cargoAsset'
   | 'cargoToken'
-  | 'cargoVendor'
-  | 'cargoAsset';
+  | 'cargoVendor';
 
 type ContractObject = {
   name: ContractNames;
@@ -51,6 +54,16 @@ const REQUEST_URLS: { [N in TNetwork]: string } = {
   development: 'https://dev-api.cargo.engineering',
   production: 'https://api.cargo.build',
 };
+
+export type ResponseType<D> =
+  | {
+      err: true;
+      data: void;
+    }
+  | {
+      err: false;
+      data: D;
+    };
 
 export type Contracts = { [Name in ContractNames]: ContractObject };
 
@@ -168,11 +181,36 @@ class Cargo extends Emitter {
     this.initialized = true;
   };
 
+  // DEPRECATED use createCargoTokenV1Instance
   public createCargoTokenInstance = (address: string) => {
     if (!this.enabled) {
       throw new Error('Enable Cargo before calling this method');
     } else {
       return this.web3.eth.contract(this.contracts.cargoToken.abi).at(address);
+    }
+  };
+
+  createCargoTokenV1Instance = this.createCargoTokenInstance;
+
+  supportsBatchMint = (address: string) => {
+    const token = this.createCargoTokenV1Instance(address);
+    return new Promise(async resolve => {
+      const supportsBatchMint = token.supportsInterface.call(
+        '0x3fa8f388',
+        (_err: any, data: boolean) => {
+          resolve(data);
+        },
+      );
+    });
+  };
+
+  createCargoTokenV2Instance = (address: string) => {
+    if (!this.enabled) {
+      throw new Error('Enable Cargo before calling this method');
+    } else {
+      return this.web3.eth
+        .contract(this.contracts.cargoBatchMint.abi)
+        .at(address);
     }
   };
 
@@ -185,27 +223,25 @@ class Cargo extends Emitter {
     fetch(`${!rawUrl ? `${this.requestUrl}${path}` : path}`, {
       cache: 'no-cache',
       ...options,
-    })
-      .then(async res => {
-        if (isJson) {
-          const json = await res.json();
-          if (res.ok) {
-            return {
-              err: false,
-              data: json,
-            };
-          }
-          return {
-            err: true,
-            data: json,
-          };
-        } else if (res.ok) {
+    }).then(async res => {
+      if (isJson) {
+        const json = await res.json();
+        if (res.ok) {
           return {
             err: false,
+            data: json,
           };
         }
-      })
-      .then(j => j);
+        return {
+          err: true,
+          data: json,
+        };
+      } else if (res.ok) {
+        return {
+          err: false,
+        };
+      }
+    });
 
   enabled: boolean = false;